Release 0.1.0: Initial working version with dog/cat API integration and quotes
